Clean up. Simplify config.
