+++ src/integrationTest/java/uk/gov/hmcts/reform/dataextractor/QueryExecutorTest.java
@@ -14,9 +14,10 @@ public class QueryExecutorTest extends DbTest {
     @Test
     public void whenSimpleSelectQueryExecuted_thenResultSetReturned() throws Exception {
         String sql = "SELECT ID, NAME FROM parent WHERE ID = 1";
-        try (QueryExecutor executor = new QueryExecutor(jdbcUrl, username, password, sql)) {
-            ResultSet resultSet = executor.execute();
-            resultSet.next();
+        try (QueryExecutor executor = new QueryExecutor(jdbcUrl, username, password, sql);
+             ResultSet resultSet = executor.execute()) {
+            boolean next = resultSet.next();
+            assertEquals(true, next);
             int result = resultSet.getInt(1);
             assertEquals(1, result);
         }
